backend/requirements.txt
# MyHealthFirstAI Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0

# Google Gemini AI
google-generativeai>=0.8.0

# Hugging Face AI (Vision backup)
huggingface-hub>=0.20.0

# Clarifai AI (Vision backup)
clarifai>=9.10.0

# Audio Processing (Voice Logging)
pydub>=0.25.1

# Image Processing
Pillow>=10.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# Monitoring & Logging
loguru>=0.7.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Development
python-dotenv>=1.0.0
